+++ cli/tests/test_global_opts.rs
@@ -705,14 +705,6 @@ fn test_config_args() {
     test.key1 = "arg1"
     [EOF]
     "#);
-    let output = list_config(&["--config-toml=test.key1='arg1'"]);
-    insta::assert_snapshot!(output, @r"
-    test.key1 = 'arg1'
-    [EOF]
-    ------- stderr -------
-    Warning: --config-toml is deprecated; use --config or --config-file instead.
-    [EOF]
-    ");
     let output = list_config(&["--config-file=file1.toml"]);
     insta::assert_snapshot!(output, @r"
     test.key1 = 'file1'
@@ -736,29 +728,15 @@ fn test_config_args() {
     let output = list_config(&[
         "--config=test.key1=arg1",
         "--config-file=file1.toml",
-        "--config-toml=test.key2='arg3'",
         "--config-file=file2.toml",
     ]);
     insta::assert_snapshot!(output, @r##"
     # test.key1 = "arg1"
     test.key1 = 'file1'
-    # test.key2 = 'file1'
-    test.key2 = 'arg3'
+    test.key2 = 'file1'
     test.key3 = 'file2'
     [EOF]
-    ------- stderr -------
-    Warning: --config-toml is deprecated; use --config or --config-file instead.
-    [EOF]
     "##);
-
-    let output = test_env.run_jj_in(".", ["config", "list", "foo", "--config-toml=foo='bar'"]);
-    insta::assert_snapshot!(output, @r"
-    foo = 'bar'
-    [EOF]
-    ------- stderr -------
-    Warning: --config-toml is deprecated; use --config or --config-file instead.
-    [EOF]
-    ");
 
     let output = test_env.run_jj_in(".", ["config", "list", "--config=foo"]);
     insta::assert_snapshot!(output, @r"